Took care of the deprecated Calendar Stuff?
Maybe? We’ll have to see.

+++ Commutable/Commutable/CommutableFirstViewController.m
@@ -205,7 +205,7 @@
     }
     else {
         cameFromLocalNotification = YES;
-        currentPage = commuteIndex;
+        currentPage = (int)commuteIndex;
         return commuteIndex;
     }
 }
